Create ContactTagIntegrationManager class
The class integrates the functionality of Person and Tag together by
implementing operations that affect both. AddressBook is used as the
proxy manager for Person and Tag-Person map. TagTree is used to keep
track of only Tag to Tag relationships. This class then supports
operations that makes changes to both.